requirements.txt
rich==13.7.1
Pillow==10.3.0
PyPDF2==3.0.1